[refactor] : Refactoring search page
